Initial commit
The marc2md script is in a good, usable state and implements most
of what I'm looking for. It extracts the key data from a MARC
record and outputs them as Markdown for use in a static blog. There's
obviously a lot more that can be done but I don't want to go down
a rabbit hole.